Factor out ArtistRoles component
