+++ root/search/components/EventResults.js
@@ -8,14 +8,12 @@
  */
 
 import React from 'react';
-import uniq from 'lodash/uniq';
 
 import {withCatalystContext} from '../../context';
 import {l} from '../../static/scripts/common/i18n';
 import {lp_attributes} from '../../static/scripts/common/i18n/attributes';
-import commaOnlyList from '../../static/scripts/common/i18n/commaOnlyList';
+import ArtistRoles from '../../static/scripts/common/components/ArtistRoles';
 import DescriptiveLink from '../../static/scripts/common/components/DescriptiveLink';
-import EntityLink from '../../static/scripts/common/components/EntityLink';
 import formatDatePeriod from '../../static/scripts/common/utility/formatDatePeriod';
 import loopParity from '../../utility/loopParity';
 import type {ResultsPropsT} from '../types';
@@ -34,20 +32,7 @@ function buildResult(result, index) {
       </td>
       <td>{event.typeName ? lp_attributes(event.typeName, 'event_type') : null}</td>
       <td>
-        <ul>
-          {event.performers.map(r =>(
-            <li key={r.entity.id}>
-              {l('{artist} ({roles})', {
-                __react: true,
-                artist: <EntityLink entity={r.entity} />,
-                roles: r.roles.length > 1
-                  // $FlowFixMe
-                  ? commaOnlyList(uniq(r.roles))
-                  : r.roles[0],
-              })}
-            </li>
-          ))}
-        </ul>
+        <ArtistRoles relations={event.performers} />
       </td>
       <td>
         <ul>
